requirements.txt
numpy
scipy
sympy
pandas
openpyxl
xlrd3
nltk
spacy
gensim
wordninja
python_Levenshtein
diff-match-patch
fuzzywuzzy
gibberish-detector
pywordsegment
scikit-learn
statsmodels
holidays
phonenumbers 
jupyterlab
ipykernel 
ipython
jupyterlab-widgets
ipyleaflet
ipywidgets
aie-ipyleaflet
matplotlib
plotly
seaborn
altair
bokeh
pygal
geoplotlib
opencv-python-headless
scikit-image
Pillow
deskew
itk
vtk
pytesseract
keras-ocr
easyocr
paddleocr
paddlepaddle
ocrmypdf
camelot-py
tabula-py
tabulate
pypdf
PyMuPDF
pdfplumber
img2pdf
pdf2image
poppler-utils
pdf2docx
fpdf2
pikepdf
ghostscript
reportlab
comtypes
outlook-msg
extract-msg
pyzbar
qreader
rasterio
rasterstats
geopandas
Fiona
psutil
tqdm
cython
pycreds
unidecode
xmltodict3
orjson
simplejson
configparser
loguru
setuptools-git
pyinstaller
Flask
flask-helmet
Django
django-cors-headers
djangorestframework
djangorestframework-jsonapi
django-filter
django-debug-toolbar
drf-yasg2
whitenoise
waitress
tornado
celery
fastapi
uvicorn
gunicorn
mysql-connector-python
pymongo
pyodbc
psycopg2
docker
PySide6
pybind11
transformers
pycocotools
albumentations
torch
torchvision
torchaudio
tensorboard
tensorflow
tensorflow-addons
tensorflow-datasets
tensorflow-estimator
tensorflow-hub
tensorflow-metadata
tensorflow_similarity
keras-tuner
GDAL
shapely